requirements.txt
Flask
Flask-Migrate
Flask-SQLAlchemy
SQLAlchemy
psycopg2-binary
requests
gunicorn
python-dotenv
